extract common component
